if you call size() on a std::vector, it returns an _unsigned_ int so need an unsigned iterator or a bigger type with a cast
git-svn-id: https://bzflag.svn.sourceforge.net/svnroot/bzflag/trunk/bzwgen@14746 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ Mesh.cxx
@@ -68,7 +68,7 @@ void Mesh::output(Output& out) {
   for (size_t i = 0; i < tc.size(); i++) out.texcoord(tc[i]);
   if (passable) out.line("  passable"); 
   if (noradar) out.line("  noradar"); 
-  for (int i = 0; i < f.size(); i++) out.face(f[i],mat);
+  for (size_t i = 0; i < f.size(); i++) out.face(f[i],mat);
   out.line("end\n");
 }
